starting branch to dev testing & make my nodejs api with pure functions
