+++ server/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "start": "node index.js",
     "lint": "eslint --fix index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha --timeout 10000"
   },
   "author": "",
   "license": "ISC",
@@ -17,8 +17,11 @@
     "node-fetch": "^2.6.1"
   },
   "devDependencies": {
+    "chai": "^4.2.0",
+    "chai-http": "^4.3.0",
     "eslint": "^7.19.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-react": "^7.22.0"
+    "eslint-plugin-react": "^7.22.0",
+    "mocha": "^8.2.1"
   }
 }
